Move SafeIndex and IndexSettings to own files
